batch_processing/raygun_ip_processing/requirements.txt
aiobotocore==2.13.1
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
attrs==23.2.0
botocore==1.34.131
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
frozenlist==1.4.1
fsspec==2024.6.0
future==1.0.0
idna==3.7
jmespath==1.0.1
minio==7.2.7
multidict==6.0.5
py4j==0.10.9.7
pycparser==2.22
pycryptodome==3.20.0
PyHive==0.7.0
pyspark==3.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
s3fs==2024.6.0
six==1.16.0
trino==0.328.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
wrapt==1.16.0
yarl==1.9.4
